normalize buffer after mixing

+++ IGenerator.cs
@@ -2,6 +2,6 @@ using System;
 
 namespace chirpcore {
     public interface IGenerator {
-        void Fill(short[] buffer, double frequency);
+        void Fill(double[] buffer, double frequency);
     }
 }
